Refactor: replaced <a> with <link> and changed file names
