+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { Header } from "./header/header"
 export { Footer } from "./footer/footer"
+export { Badge } from "./miscellaneous/miscellaneous";
